feat: eslintで怒られた。halfSizeCardsを別のとこで使ってたらファイル分けて定義しろとのことです
