+++ src/constants.ts
@@ -1,0 +1,7 @@
+// src/constants.ts
+export const halfSizeCards: string[] = [
+    "Top languages used in repository card",
+    "Top languages in commits card",
+    "GitHub stats card",
+    "Productive time card",
+  ];
